refactor(describegpt): extract session cluster into submodule
Move the chat-mode session cluster into src/cmd/describegpt/session.rs:
SessionMessage / SessionState plus normalize_session_path, load_session,
save_session (with the atomic-write fix preserved), track_sql_error_in_session,
update_session_after_sql_success, apply_sliding_window, generate_summary,
and check_message_relevance.

These functions form a cohesive unit around reading / writing the markdown
session file, sliding-window compaction, and LLM-assisted summarization /
relevance checks. Couplings back to the parent (check_model, get_completion,
extract_sql_sample, PromptType, Args) are imported via super::.

No behavior change. Main file drops from 5,062 to 4,622 lines — cumulative
reduction across this refactor branch: 954 lines (17%).
Clippy clean; all 64 integration + 7 unit tests pass.
