+++ src/cmd/describegpt.rs
@@ -384,10 +384,16 @@ use crate::{lookup, lookup::LookupTableOptions};
 
 mod dictionary;
 mod formatters;
+mod session;
 
 use dictionary::{
     combine_dictionary_entries, generate_code_based_dictionary, parse_frequency_csv,
     parse_llm_dictionary_response, parse_stats_csv,
+};
+use session::{
+    SessionMessage, SessionState, apply_sliding_window, check_message_relevance, load_session,
+    normalize_session_path, save_session, track_sql_error_in_session,
+    update_session_after_sql_success,
 };
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, EnumString, Display)]
@@ -464,23 +470,6 @@ struct Args {
     flag_no_score_sql:      bool,
     flag_score_threshold:   u32,
     flag_score_max_retries: u32,
-}
-
-#[derive(Debug, Clone)]
-struct SessionMessage {
-    role:      String,
-    content:   String,
-    #[allow(dead_code)]
-    timestamp: String,
-}
-
-#[derive(Debug, Clone)]
-struct SessionState {
-    baseline_sql: Option<String>,
-    messages:     Vec<SessionMessage>,
-    sql_results:  Option<String>,
-    sql_errors:   Vec<String>,
-    summary:      Option<String>,
 }
 
 #[derive(Debug, Deserialize)]
@@ -3521,43 +3510,6 @@ fn remove_prompt_cache_entries_for_both_flags(
     }
 }
 
-/// Track SQL error in session state if present
-fn track_sql_error_in_session(
-    session_state: Option<&mut SessionState>,
-    normalized_session_path: Option<&String>,
-    error_msg: String,
-) {
-    if let Some(state) = session_state {
-        state.sql_errors.push(error_msg);
-        if let Some(normalized_path) = normalized_session_path {
-            let _ = save_session(Path::new(normalized_path), state);
-        }
-    }
-}
-
-/// Update session state after successful SQL execution
-fn update_session_after_sql_success(
-    session_state: Option<&mut SessionState>,
-    sql_results: &str,
-    sql_query: &str,
-) {
-    if let Some(state) = session_state {
-        let results_path = Path::new(sql_results).with_extension("csv");
-        if results_path.exists()
-            && let Ok(sample) = extract_sql_sample(&results_path)
-        {
-            state.sql_results = Some(sample);
-            state.sql_errors.clear(); // Clear errors on success
-        }
-
-        // Extract and store baseline SQL only after successful execution
-        // This ensures baseline SQL is only set when the query executes successfully
-        if state.baseline_sql.is_none() {
-            state.baseline_sql = Some(sql_query.to_string());
-        }
-    }
-}
-
 #[allow(dead_code)]
 /// Helper function to handle SQL error cases by invalidating cache and saving the failed query
 fn handle_sql_error(
@@ -3577,268 +3529,6 @@ fn handle_sql_error(
         return fail_clierror!("Failed to copy SQL query to {sql_results_path:?}: {e}");
     }
     fail_clierror!("{error_msg}")
-}
-
-/// Normalize session path to always have .md extension
-fn normalize_session_path(session_path: &str) -> String {
-    let path = Path::new(session_path);
-    if let Some(ext) = path.extension()
-        && ext == "md"
-    {
-        return session_path.to_string();
-    }
-    // If no extension or wrong extension, ensure .md extension
-    // Use with_extension which replaces existing extension or adds if none exists
-    path.with_extension("md").to_string_lossy().to_string()
-}
-
-/// Load session state from a markdown file
-fn load_session(session_path: &Path) -> CliResult<SessionState> {
-    if !session_path.exists() {
-        return Ok(SessionState {
-            baseline_sql: None,
-            messages:     Vec::new(),
-            sql_results:  None,
-            sql_errors:   Vec::new(),
-            summary:      None,
-        });
-    }
-
-    let content = fs::read_to_string(session_path)?;
-    let mut state = SessionState {
-        baseline_sql: None,
-        messages:     Vec::new(),
-        sql_results:  None,
-        sql_errors:   Vec::new(),
-        summary:      None,
-    };
-
-    let lines: Vec<&str> = content.lines().collect();
-    let mut i = 0;
-    let mut current_content = String::new();
-    let mut current_role = String::new();
-
-    while i < lines.len() {
-        let line = lines[i].trim();
-
-        if line.starts_with("# Session:") {
-            // Skip header
-        } else if line == "## Baseline SQL Query" {
-            current_content.clear();
-            i += 1;
-            // Skip opening ```sql
-            if i < lines.len() && lines[i].trim() == "```sql" {
-                i += 1;
-            }
-            // Read SQL until closing ```
-            while i < lines.len() && !lines[i].trim().starts_with("```") {
-                if !current_content.is_empty() {
-                    current_content.push('\n');
-                }
-                current_content.push_str(lines[i]);
-                i += 1;
-            }
-            state.baseline_sql = Some(current_content.trim().to_string());
-            current_content.clear();
-        } else if line == "## Conversation History" {
-            i += 1;
-            // Parse messages
-            let mut in_content_section = false;
-            let mut in_code_block = false;
-            while i < lines.len() {
-                let msg_line = lines[i];
-                let msg_line_trimmed = msg_line.trim();
-
-                // Check if we've hit the next section header (## at start of line, not indented)
-                // Only break if we're not inside a code block
-                if !in_code_block
-                    && msg_line_trimmed.starts_with("##")
-                    && !msg_line_trimmed.starts_with("###")
-                {
-                    // Save the current message before breaking
-                    if !current_content.is_empty() && !current_role.is_empty() {
-                        state.messages.push(SessionMessage {
-                            role:      current_role.clone(),
-                            content:   current_content.trim().to_string(),
-                            timestamp: chrono::Utc::now().to_rfc3339(),
-                        });
-                    }
-                    break;
-                }
-
-                // Track code blocks to avoid breaking on ## inside them
-                if msg_line_trimmed.starts_with("```") {
-                    in_code_block = !in_code_block;
-                }
-
-                if msg_line_trimmed.starts_with("### Message") {
-                    // New message - save previous if exists
-                    if !current_content.is_empty() && !current_role.is_empty() {
-                        state.messages.push(SessionMessage {
-                            role:      current_role.clone(),
-                            content:   current_content.trim().to_string(),
-                            timestamp: chrono::Utc::now().to_rfc3339(),
-                        });
-                    }
-                    // Reset for new message
-                    current_content.clear();
-                    current_role.clear();
-                    in_content_section = false;
-                    in_code_block = false;
-                } else if msg_line_trimmed.starts_with("**Role:**") {
-                    current_role = msg_line_trimmed.replace("**Role:**", "").trim().to_string();
-                    in_content_section = false;
-                } else if msg_line_trimmed.starts_with("**Content:**") {
-                    // Content section starts - clear any previous content for this message
-                    current_content.clear();
-                    in_content_section = true;
-                } else if in_content_section {
-                    // We're in the content section - add everything (including empty lines and code
-                    // blocks) Always add a newline before adding content
-                    // (except for the very first line)
-                    if !current_content.is_empty() {
-                        current_content.push('\n');
-                    }
-                    // Add the line as-is (preserving original formatting)
-                    current_content.push_str(msg_line);
-                }
-                i += 1;
-            }
-            // Add last message if we didn't break on a section header
-            if !current_content.is_empty() && !current_role.is_empty() {
-                state.messages.push(SessionMessage {
-                    role:      current_role.clone(),
-                    content:   current_content.trim().to_string(),
-                    timestamp: chrono::Utc::now().to_rfc3339(),
-                });
-            }
-            continue;
-        } else if line == "## SQL Results (Last Successful)" {
-            current_content.clear();
-            i += 1;
-            // Skip opening ```csv
-            if i < lines.len() && lines[i].trim() == "```csv" {
-                i += 1;
-            }
-            // Read CSV until closing ```
-            while i < lines.len() && !lines[i].trim().starts_with("```") {
-                if !current_content.is_empty() {
-                    current_content.push('\n');
-                }
-                current_content.push_str(lines[i]);
-                i += 1;
-            }
-            state.sql_results = Some(current_content.trim().to_string());
-            current_content.clear();
-        } else if line == "## SQL Errors" {
-            i += 1;
-            // Read error list items
-            while i < lines.len() && !lines[i].trim().starts_with("##") {
-                let line = lines[i].trim();
-                if line.starts_with("- ") {
-                    let error = line.strip_prefix("- ").unwrap_or(line).to_string();
-                    state.sql_errors.push(error);
-                }
-                i += 1;
-            }
-            continue;
-        } else if line == "## Summary" {
-            current_content.clear();
-            i += 1;
-            // Read summary until next section or end
-            while i < lines.len() && !lines[i].trim().starts_with("##") {
-                if !current_content.is_empty() {
-                    current_content.push('\n');
-                }
-                current_content.push_str(lines[i]);
-                i += 1;
-            }
-            state.summary = Some(current_content.trim().to_string());
-            current_content.clear();
-            continue;
-        }
-        i += 1;
-    }
-
-    Ok(state)
-}
-
-/// Save session state to a markdown file.
-///
-/// Writes atomically via a temp file in the same directory, then renames into place,
-/// so a crash mid-write leaves either the previous good file or the new one — never
-/// a truncated file.
-fn save_session(session_path: &Path, state: &SessionState) -> CliResult<()> {
-    use std::{fmt::Write as _, io::Write as _}; // import without risk of name clashing
-
-    // Ensure parent directory exists
-    let parent_dir = session_path.parent().filter(|p| !p.as_os_str().is_empty());
-    if let Some(parent) = parent_dir {
-        fs::create_dir_all(parent)?;
-    }
-
-    let mut content = String::new();
-    let _ = write!(
-        content,
-        "# Session: {}\n\n",
-        session_path
-            .file_name()
-            .unwrap_or_default()
-            .to_string_lossy()
-    );
-
-    // Baseline SQL Query
-    if let Some(ref sql) = state.baseline_sql {
-        content.push_str("## Baseline SQL Query\n\n");
-        content.push_str("```sql\n");
-        content.push_str(sql);
-        content.push_str("\n```\n\n");
-    }
-
-    // Conversation History
-    content.push_str("## Conversation History\n\n");
-    for (idx, msg) in state.messages.iter().enumerate() {
-        let _ = write!(content, "### Message {}\n\n", idx + 1);
-        let _ = write!(content, "**Role:** {}\n\n", msg.role);
-        let _ = write!(content, "**Content:**\n\n{}\n\n", msg.content);
-    }
-
-    // SQL Results
-    if let Some(ref results) = state.sql_results {
-        content.push_str("## SQL Results (Last Successful)\n\n");
-        content.push_str("```csv\n");
-        content.push_str(results);
-        content.push_str("\n```\n\n");
-    }
-
-    // SQL Errors
-    if !state.sql_errors.is_empty() {
-        content.push_str("## SQL Errors\n\n");
-        for error in &state.sql_errors {
-            let _ = writeln!(content, "- {error}");
-        }
-        content.push('\n');
-    }
-
-    // Summary
-    if let Some(ref summary) = state.summary {
-        content.push_str("## Summary\n\n");
-        content.push_str(summary);
-        content.push('\n');
-    }
-
-    // Atomic write: create temp file in the destination directory (so the final
-    // rename is a same-filesystem op), write+flush, then persist over the target.
-    let tmp_dir = parent_dir.unwrap_or(Path::new("."));
-    let mut tmp = tempfile::Builder::new()
-        .prefix(".describegpt-session-")
-        .suffix(".tmp")
-        .tempfile_in(tmp_dir)?;
-    tmp.as_file_mut().write_all(content.as_bytes())?;
-    tmp.as_file_mut().sync_all()?;
-    tmp.persist(session_path)
-        .map_err(|e| CliError::from(e.error))?;
-    Ok(())
 }
 
 /// Extract first 10 rows from a CSV file
@@ -3867,136 +3557,6 @@ fn extract_sql_sample(csv_path: &Path) -> CliResult<String> {
     }
 
     Ok(result.trim().to_string())
-}
-
-/// Generate summary of old messages using LLM
-fn generate_summary(
-    old_messages: &[SessionMessage],
-    args: &Args,
-    client: &Client,
-    api_key: &str,
-) -> CliResult<String> {
-    use std::fmt::Write as _; // import without risk of name clashing
-
-    let mut summary_prompt = String::from(
-        "Please provide a concise summary of the following conversation history. Focus on the key \
-         SQL query refinements, user requests, and assistant responses:\n\n",
-    );
-
-    for msg in old_messages {
-        let _ = write!(summary_prompt, "{}: {}\n\n", msg.role, msg.content);
-    }
-
-    summary_prompt
-        .push_str("\nProvide a brief summary that captures the essence of this conversation:");
-
-    let system_prompt = "You are a helpful assistant that summarizes conversation history for SQL \
-                         query refinement sessions.";
-    let messages = json!([
-        {"role": "system", "content": system_prompt},
-        {"role": "user", "content": summary_prompt}
-    ]);
-
-    let model = check_model(client, Some(api_key), args)?;
-    let completion = get_completion(args, client, &model, api_key, &messages, PromptType::Prompt)?;
-    Ok(completion.response)
-}
-
-/// Apply sliding window to session messages, summarizing older ones if needed
-fn apply_sliding_window(
-    state: &mut SessionState,
-    max_len: usize,
-    args: &Args,
-    client: &Client,
-    api_key: &str,
-) -> CliResult<()> {
-    if state.messages.len() <= max_len {
-        return Ok(());
-    }
-
-    let num_to_summarize = state.messages.len() - max_len;
-    let old_messages: Vec<SessionMessage> = state.messages.drain(..num_to_summarize).collect();
-
-    // Generate summary of old messages
-    let summary = generate_summary(&old_messages, args, client, api_key)?;
-
-    // Combine with existing summary if present
-    if let Some(ref existing_summary) = state.summary {
-        state.summary = Some(format!("{existing_summary}\n\n{summary}"));
-    } else {
-        state.summary = Some(summary);
-    }
-
-    Ok(())
-}
-
-/// Check if a message is relevant to the baseline SQL query
-fn check_message_relevance(
-    prompt: &str,
-    baseline_sql: &str,
-    args: &Args,
-    client: &Client,
-    api_key: &str,
-) -> CliResult<bool> {
-    // Heuristic check: Look for SQL-related keywords
-    let sql_keywords = [
-        "sql",
-        "query",
-        "select",
-        "where",
-        "join",
-        "group",
-        "order",
-        "filter",
-        "refine",
-        "modify",
-        "change",
-        "update",
-        "fix",
-        "correct",
-        "improve",
-        "add",
-        "remove",
-        "include",
-        "exclude",
-        "sort",
-        "aggregate",
-        "count",
-    ];
-
-    let prompt_lower = prompt.to_lowercase();
-    let has_sql_keywords = sql_keywords.iter().any(|kw| prompt_lower.contains(kw));
-
-    // Also check if prompt references previous query/results
-    let has_references = prompt_lower.contains("previous")
-        || prompt_lower.contains("last")
-        || prompt_lower.contains("above")
-        || prompt_lower.contains("before");
-
-    if has_sql_keywords || has_references {
-        return Ok(true);
-    }
-
-    // LLM check: Ask LLM if message is related to refining the SQL query
-    let relevance_prompt = format!(
-        "The user has been working on refining a SQL query. The baseline SQL query \
-         is:\n\n```sql\n{baseline_sql}\n```\n\nUser's new message: \"{prompt}\"\n\nIs this \
-         message related to refining, modifying, or improving the SQL query above? Answer with \
-         only 'yes' or 'no'."
-    );
-
-    let system_prompt = "You are a helpful assistant that determines if user messages are related \
-                         to SQL query refinement.";
-    let messages = json!([
-        {"role": "system", "content": system_prompt},
-        {"role": "user", "content": relevance_prompt}
-    ]);
-
-    let model = check_model(client, Some(api_key), args)?;
-    let completion = get_completion(args, client, &model, api_key, &messages, PromptType::Prompt)?;
-    let response_lower = completion.response.to_lowercase().trim().to_string();
-
-    Ok(response_lower.contains("yes") || response_lower == "y")
 }
 
 /// Determine which additional columns to include based on args
